requirements.txt
diffusers
datasets
fastapi
wandb
lightning
torchvision
pandas
numpy
rich
tqdm
transformers
opencv-python-headless